// web/src/shared/api/generated/model/subjectDetailResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Elmanhg.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { UnitResult } from './unitResult';

export interface SubjectDetailResult {
  id: string;
  name: string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  order: number | string;
  units: UnitResult[];
}